requirements.txt
absl-py==2.3.1
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiomysql==0.3.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.4
asttokens==3.0.1
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
blinker==1.9.0
boto3==1.40.61
botocore==1.40.61
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.104.1
filelock==3.20.2
firebase_admin==7.1.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gspread==6.2.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mysql-connector==2.2.9
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.14.0
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.46.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typesense==1.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.246
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
wrapt==1.17.3
yarl==1.22.0
